Added a models module to hold the models for other classes
