+++ src/unofficial_tabdeal_api/order.py
@@ -1,39 +1,13 @@
 """This module holds the OrderClass."""
 # pylint: disable=R0913
 
-from decimal import Decimal
 from typing import Any
 
-from pydantic import BaseModel, Field
-
 from unofficial_tabdeal_api.base import BaseClass
-from unofficial_tabdeal_api.constants import GET_ORDERS_HISTORY_URI, MAX_ALLOWED_MARGIN_LEVEL
-from unofficial_tabdeal_api.enums import OrderSide, OrderState
+from unofficial_tabdeal_api.constants import GET_ORDERS_HISTORY_URI
+from unofficial_tabdeal_api.enums import OrderState
 from unofficial_tabdeal_api.exceptions import OrderNotFoundInSpecifiedHistoryRangeError
 from unofficial_tabdeal_api.utils import find_order_by_id
-
-
-class MarginOrder(BaseModel):
-    """This is the class storing information about a margin order."""
-
-    isolated_symbol: str
-    """Symbol of the order, e.g. BTCUSDT"""
-    order_price: Decimal = Field(..., gt=0)
-    """Price of the order, e.g. 10000.00"""
-    order_side: OrderSide
-    """Side of the order, either BUY or SELL"""
-    margin_level: Decimal = Field(..., gt=0, le=MAX_ALLOWED_MARGIN_LEVEL)
-    """Margin level of the order, e.g. 1.5"""
-    deposit_amount: Decimal = Field(..., gt=0)
-    """Deposit amount for the order, e.g. 1000.00"""
-    stop_loss_percent: Decimal = Field(..., ge=0)
-    """Percentile of tolerate-able loss, e.g. 5 for 5%"""
-    take_profit_percent: Decimal = Field(..., ge=0)
-    """Percentile of expected profit, e.g. 10 for 10%"""
-    volume_fraction_allowed: bool
-    """Whether volume fraction is allowed, e.g. True or False"""
-    volume_precision: int = 0
-    """Precision of the volume, Defaults to 0"""
 
 
 class OrderClass(BaseClass):
@@ -66,7 +40,7 @@ class OrderClass(BaseClass):
         }
 
         # We get the data from server
-        response = await self._get_data_from_server(
+        response: dict[str, Any] | list[dict[str, Any]] = await self._get_data_from_server(
             connection_url=GET_ORDERS_HISTORY_URI,
             queries=connection_query,
         )
